requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_py-xgboost-mutex=2.0=cpu_0
astroid=2.3.3=py38_1
attrs=19.3.0=py_0
backcall=0.1.0=py_0
bleach=3.1.4=pyh9f0ad1d_0
ca-certificates=2020.4.5.1=hecc5488_0
certifi=2020.4.5.1=py38h32f6830_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
entrypoints=0.3=py38h32f6830_1001
flake8=3.7.9=py38h32f6830_1
freetype=2.10.1=he06d7ca_0
icu=64.2=he1b5a44_1
importlib-metadata=1.6.0=py38h32f6830_0
importlib_metadata=1.6.0=0
ipykernel=5.2.1=py38h23f93f0_0
ipython=7.14.0=py38h32f6830_0
ipython_genutils=0.2.0=py_1
isort=4.3.21=py38h32f6830_1
jedi=0.17.0=py38h32f6830_0
jinja2=2.11.2=pyh9f0ad1d_0
joblib=0.14.1=py_0
jsonschema=3.2.0=py38h32f6830_1
jupyter_client=6.1.3=py_0
jupyter_contrib_core=0.3.3=py_2
jupyter_contrib_nbextensions=0.5.1=py38_0
jupyter_core=4.6.3=py38h32f6830_1
jupyter_highlight_selected_word=0.2.0=py38_1000
jupyter_latex_envs=1.4.6=py38_1000
jupyter_nbextensions_configurator=0.4.1=py38_0
kiwisolver=1.2.0=py38hbf85e49_0
lazy-object-proxy=1.4.3=py38h1e0a361_2
ld_impl_linux-64=2.34=h53a641e_0
libblas=3.8.0=16_openblas
libcblas=3.8.0=16_openblas
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.2.0=h24d8f2e_2
libgfortran-ng=7.3.0=hdf63c60_5
libiconv=1.15=h516909a_1006
liblapack=3.8.0=16_openblas
libopenblas=0.3.9=h5ec1e0e_0
libpng=1.6.37=hed695b0_1
libsodium=1.0.17=h516909a_0
libstdcxx-ng=9.2.0=hdf63c60_2
libxgboost=1.0.2=he1b5a44_1
libxml2=2.9.10=hee79883_0
libxslt=1.1.33=h31b3aaa_0
llvm-openmp=10.0.0=hc9558a2_0
lxml=4.5.0=py38hbb43d70_1
markupsafe=1.1.1=py38h1e0a361_1
matplotlib-base=3.2.1=py38h2af1d28_0
mccabe=0.6.1=py_1
mistune=0.8.4=py38h1e0a361_1001
nb_conda_kernels=2.2.3=py38_0
nbconvert=5.6.1=py38h32f6830_1
nbformat=5.0.6=py_0
ncurses=6.1=hf484d3e_1002
notebook=6.0.3=py38_0
numpy=1.18.1=py38h8854b6b_1
openssl=1.1.1g=h516909a_0
pandas=1.0.3=py38hcb8c335_1
pandoc=2.9.2.1=0
pandocfilters=1.4.2=py_1
parso=0.7.0=pyh9f0ad1d_0
patsy=0.5.1=py_0
pexpect=4.8.0=py38h32f6830_1
pickleshare=0.7.5=py38h32f6830_1001
pip=20.1=pyh9f0ad1d_0
prometheus_client=0.7.1=py_0
prompt-toolkit=3.0.5=py_0
ptyprocess=0.6.0=py_1001
py-xgboost=1.0.2=py38h32f6830_1
pycodestyle=2.5.0=py_0
pyflakes=2.1.1=py_0
pygments=2.6.1=py_0
pylint=2.4.4=py38_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyrsistent=0.16.0=py38h1e0a361_0
python=3.8.2=he5300dc_7_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytz=2020.1=pyh9f0ad1d_0
pyyaml=5.3.1=py38h1e0a361_0
pyzmq=19.0.0=py38ha71036d_1
readline=8.0=hf8c457e_0
scikit-learn=0.22.2.post1=py38hcdab131_0
scipy=1.4.1=py38h18bccfc_3
seaborn=0.10.1=py_0
send2trash=1.5.0=py_0
setuptools=46.1.3=py38h32f6830_0
six=1.14.0=py_1
sklearn-pandas=1.8.0=pypi_0
sqlite=3.30.1=hcee41ef_0
src-abhi8893=0.0.1=dev_0
statsmodels=0.11.1=py38h1e0a361_1
terminado=0.8.3=py38h32f6830_1
testpath=0.4.4=py_0
tk=8.6.10=hed695b0_0
tornado=6.0.4=py38h1e0a361_1
traitlets=4.3.3=py38h32f6830_1
wcwidth=0.1.9=pyh9f0ad1d_0
webencodings=0.5.1=py_1
wheel=0.34.2=py_1
wrapt=1.12.1=py38h1e0a361_1
xgboost=1.0.2=py38h950e882_1
xz=5.2.5=h516909a_0
yaml=0.2.4=h516909a_0
zeromq=4.3.2=he1b5a44_2
zipp=3.1.0=py_0
zlib=1.2.11=h516909a_1006
